AN-9: add middleware role/verifyowner, change login response to object user

+++ app/Repositories/CommentRepository.php
@@ -51,9 +51,12 @@ class CommentRepository
     {
 
         $comment = Comment::findOrFail($id);
-        $comment->delete();
-        return $comment;
-
+        $user = auth()->id();
+        if ($comment->user_id == $user) {
+            $comment->delete();
+            return $comment;
+        } else
+            return response()->json(['message' => trans('message.unauthorized')]);
     }
 
     public function listAdComments($ad_id)
